Check if there is a necessary update of the message recipients. (#211)
Add business event in case of a recipient update.
Rename / refactoring.

+++ agrirouter-middleware-business/src/main/java/de/agrirouter/middleware/business/listener/UpdateRecipientsForEndpointEventListener.java
@@ -6,14 +6,21 @@ import com.google.protobuf.InvalidProtocolBufferException;
 import de.agrirouter.middleware.api.events.UpdateRecipientsForEndpointEvent;
 import de.agrirouter.middleware.api.logging.BusinessOperationLogService;
 import de.agrirouter.middleware.api.logging.EndpointLogInformation;
+import de.agrirouter.middleware.business.cache.events.BusinessEvent;
+import de.agrirouter.middleware.business.cache.events.BusinessEventApplicationEvent;
+import de.agrirouter.middleware.business.cache.events.BusinessEventType;
+import de.agrirouter.middleware.domain.Endpoint;
 import de.agrirouter.middleware.domain.MessageRecipient;
 import de.agrirouter.middleware.persistence.EndpointRepository;
 import de.agrirouter.middleware.persistence.MessageRecipientRepository;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.context.ApplicationEventPublisher;
 import org.springframework.context.event.EventListener;
 import org.springframework.stereotype.Service;
 
+import java.time.Instant;
 import java.util.HashSet;
+import java.util.concurrent.atomic.AtomicBoolean;
 
 /**
  * Service to update the endpoint status.
@@ -26,15 +33,18 @@ public class UpdateRecipientsForEndpointEventListener {
     private final DecodeMessageService decodeMessageService;
     private final MessageRecipientRepository messageRecipientRepository;
     private final BusinessOperationLogService businessOperationLogService;
+    private final ApplicationEventPublisher applicationEventPublisher;
 
     public UpdateRecipientsForEndpointEventListener(EndpointRepository endpointRepository,
                                                     DecodeMessageService decodeMessageService,
                                                     MessageRecipientRepository messageRecipientRepository,
-                                                    BusinessOperationLogService businessOperationLogService) {
+                                                    BusinessOperationLogService businessOperationLogService,
+                                                    ApplicationEventPublisher applicationEventPublisher) {
         this.endpointRepository = endpointRepository;
         this.messageRecipientRepository = messageRecipientRepository;
         this.decodeMessageService = decodeMessageService;
         this.businessOperationLogService = businessOperationLogService;
+        this.applicationEventPublisher = applicationEventPublisher;
     }
 
     /**
@@ -49,32 +59,75 @@ public class UpdateRecipientsForEndpointEventListener {
         if (optionalEndpoint.isPresent()) {
             final var endpoint = optionalEndpoint.get();
             if (null != updateRecipientsForEndpointEvent.getFetchMessageResponse()) {
-                log.debug("Remove all of the former message recipients.");
-                messageRecipientRepository.deleteAll(endpoint.getMessageRecipients());
-                endpoint.setMessageRecipients(new HashSet<>());
                 final var fetchMessageResponse = updateRecipientsForEndpointEvent.getFetchMessageResponse();
                 final var decodedMessageResponse = decodeMessageService.decode(fetchMessageResponse.getCommand().getMessage());
-                // FIXME Replace this one after updating to the latest release.
                 final var listEndpointsResponse = Endpoints.ListEndpointsResponse.parseFrom(decodedMessageResponse.getResponsePayloadWrapper().getDetails().getValue());
-                listEndpointsResponse.getEndpointsList().forEach(e -> e.getMessageTypesList().forEach(messageType -> {
-                    final var messageRecipient = new MessageRecipient();
-                    messageRecipient.setAgrirouterEndpointId(e.getEndpointId());
-                    messageRecipient.setEndpointName(e.getEndpointName());
-                    messageRecipient.setEndpointType(e.getEndpointType());
-                    messageRecipient.setExternalId(e.getExternalId());
-                    messageRecipient.setTechnicalMessageType(messageType.getTechnicalMessageType());
-                    messageRecipient.setDirection(messageType.getDirection().name());
-                    messageRecipientRepository.save(messageRecipient);
-                    endpoint.getMessageRecipients().add(messageRecipient);
-                }));
-                log.debug("There were {} recipients found for the endpoint '{}'.", endpoint.getMessageRecipients().size(), endpoint.getExternalEndpointId());
-                log.trace("{}", endpoint.getMessageRecipients());
-                endpointRepository.save(endpoint);
-                businessOperationLogService.log(new EndpointLogInformation(endpoint.getExternalEndpointId(), endpoint.getAgrirouterEndpointId()), "Recipients updated.");
+                if (checkIfAnUpdateOfTheRecipientsIsNeeded(endpoint, listEndpointsResponse)) {
+                    log.debug("Remove all of the former message recipients.");
+                    messageRecipientRepository.deleteAll(endpoint.getMessageRecipients());
+                    endpoint.setMessageRecipients(new HashSet<>());
+                    listEndpointsResponse.getEndpointsList().forEach(e -> e.getMessageTypesList().forEach(messageType -> {
+                        final var messageRecipient = new MessageRecipient();
+                        messageRecipient.setAgrirouterEndpointId(e.getEndpointId());
+                        messageRecipient.setEndpointName(e.getEndpointName());
+                        messageRecipient.setEndpointType(e.getEndpointType());
+                        messageRecipient.setExternalId(e.getExternalId());
+                        messageRecipient.setTechnicalMessageType(messageType.getTechnicalMessageType());
+                        messageRecipient.setDirection(messageType.getDirection().name());
+                        messageRecipientRepository.save(messageRecipient);
+                        endpoint.getMessageRecipients().add(messageRecipient);
+                    }));
+                    log.debug("There were {} recipients found for the endpoint '{}'.", endpoint.getMessageRecipients().size(), endpoint.getExternalEndpointId());
+                    log.trace("{}", endpoint.getMessageRecipients());
+                    endpointRepository.save(endpoint);
+                    businessOperationLogService.log(new EndpointLogInformation(endpoint.getExternalEndpointId(), endpoint.getAgrirouterEndpointId()), "Recipients updated.");
+                    applicationEventPublisher.publishEvent(new BusinessEventApplicationEvent(this, endpoint.getExternalEndpointId(), new BusinessEvent(Instant.now(), BusinessEventType.RECIPIENTS_UPDATED)));
+                }
+            } else {
+                log.warn("There is no change within the recipients of the endpoint. Therefore, the recipients were not updated.");
             }
         } else {
             log.warn("The endpoint was not found in the database, the message was deleted but not saved.");
         }
     }
+
+    private boolean checkIfAnUpdateOfTheRecipientsIsNeeded(Endpoint endpoint, Endpoints.ListEndpointsResponse listEndpointsResponse) {
+        AtomicBoolean updateNeeded = new AtomicBoolean(false);
+        for (Endpoints.ListEndpointsResponse.Endpoint e : listEndpointsResponse.getEndpointsList()) {
+            if (updateNeeded.get()) {
+                break;
+            }
+            for (Endpoints.ListEndpointsResponse.MessageType mt : e.getMessageTypesList()) {
+                var thereAMessageRecipientWithTheSameValues = isThereAMessageRecipientWithTheSameValues(endpoint, e, mt);
+                if (!thereAMessageRecipientWithTheSameValues) {
+                    updateNeeded.set(true);
+                    break;
+                }
+            }
+        }
+        return updateNeeded.get();
+    }
+
+    private boolean isThereAMessageRecipientWithTheSameValues(Endpoint endpoint, Endpoints.ListEndpointsResponse.Endpoint e, Endpoints.ListEndpointsResponse.MessageType mt) {
+        var thereIsAMessageRecipientWithTheSameValues = new AtomicBoolean(false);
+        endpoint.getMessageRecipients().stream().filter(mr -> compareTheMessageRecipientAttributes(mr, e, mt)).findFirst().ifPresentOrElse(
+                messageRecipient -> {
+                    log.debug("The recipient '{}' is already known.", messageRecipient.getAgrirouterEndpointId());
+                    thereIsAMessageRecipientWithTheSameValues.set(true);
+                },
+                () -> log.debug("The recipient '{}' is not known yet.", e.getEndpointId())
+        );
+        return thereIsAMessageRecipientWithTheSameValues.get();
+    }
+
+    private boolean compareTheMessageRecipientAttributes(MessageRecipient mr, Endpoints.ListEndpointsResponse.Endpoint e, Endpoints.ListEndpointsResponse.MessageType mt) {
+        return mr.getAgrirouterEndpointId().equals(e.getEndpointId()) &&
+                mr.getEndpointName().equals(e.getEndpointName()) &&
+                mr.getEndpointType().equals(e.getEndpointType()) &&
+                mr.getExternalId().equals(e.getExternalId()) &&
+                mr.getTechnicalMessageType().equals(mt.getTechnicalMessageType()) &&
+                mr.getDirection().equals(mt.getDirection().name());
+    }
+
 
 }
